fix: 타입 분리
